# SWAGGER/BankingAPISwagger.yaml
openapi: 3.0.0
info:
  title: Banking API
  description: Banking API Group 1 Endpoints
  version: "1.0.0"
  contact:
    email: 683029@student.inhollan.nl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/JOANIDEVELOP/BankingAPIProject/1.0.0
paths:
...
  /transactions/transfers:
    post:
      tags:
        - Customer
        - Employee
      summary: Transfer funds to other customers' accounts
      operationId: createTransferToOtherCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Transfer successfully completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid transfer details, incomplete or invalid data.
        '404':
          description: Account not found (either source or destination).
        '422':
          description: Transfer cannot be completed, insufficient funds or account restrictions.
      security:
        - bearerAuth: []
...
  /transactions/{userId}/history:
    get:
      tags:
        - Customer
      summary: Get user transaction history
      operationId: getUserTransactionHistory
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction history retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '404':
          description: User ID not found.
      security:
        - bearerAuth: []
...
  /transactions/search:
    get:
      tags:
        - Customer
      summary: Search and filter transactions
      operationId: searchTransactions
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
        - in: query
          name: amountLessThan
          schema:
            type: number
          description: Filter transactions less than a specific amount
        - in: query
          name: amountGreaterThan
          schema:
            type: number
          description: Filter transactions greater than a specific amount
        - in: query
          name: amountEqualTo
          schema:
            type: number
          description: Filter transactions equal to a specific amount
        - in: query
          name: IBAN
          schema:
            type: string
      responses:
        '200':
          description: Transactions found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid query parameters, such as improperly formatted dates or invalid numeric filters.
        '404':
          description: No transactions found matching the criteria.

      security:
        - bearerAuth: []
...
  /transactions:
    get:
      tags:
        - Employee
      summary: Get all transactions
      description: This can only be done by the employee.
      operationId: getAllTransactions
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/ATM:
    get:
      tags:
        - Employee
      summary: Get all transactions from the ATM
      description: This can only be done by the employee.
      operationId: getATMTransactions
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/ATM/{userId}:
    get:
      tags:
        - Employee
      summary: Get all transactions from the ATM by user ID
      description: This can only be done by the employee.
      operationId: getATMTransactionsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of user which initialized the transactions
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '404':
          description: User not found
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/ATM/withdrawals/{userId}:
    get:
      tags:
        - Employee
      summary: Get all withdrawals from the ATM by user ID
      description: This can only be done by the employee.
      operationId: getATMWithdrawalsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of user which initialized the transactions
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '404':
          description: User not found
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/ATM/deposits/{userId}:
    get:
      tags:
        - Employee
      summary: Get all deposits from the ATM by user ID
      description: This can only be done by the employee.
      operationId: getATMDepositsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of user which initialized the transactions
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '404':
          description: User not found
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []   
  /transactions/byCustomers:
    get:
      tags:
        - Employee
      summary: Get all transactions initialized by customers
      description: This can only be done by the employee.
      operationId: getTransactionsInitializedByCustomers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/byEmployees:
    get:
      tags:
        - Employee
      summary: Get all transactions initialized by employees
      description: This can only be done by the employee.
      operationId: getTransactionsInitializedByEmployees
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/online:
    get:
      tags:
        - Employee
      summary: Get all transactions made online
      description: This can only be done by the employee.
      operationId: getOnlineTransactions
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/online/byCustomers:
    get:
      tags:
        - Employee
      summary: Get all transactions made online by the customers
      description: This can only be done by the employee.
      operationId: getOnlineTransactionsByCustomers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/online/byEmployees:
    get:
      tags:
        - Employee
      summary: Get all transactions made online by the employees
      description: This can only be done by the employee.
      operationId: getOnlineTransactionsByEmployees
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
        "400":
          description: Invalid status value.
      security:
        - bearerAuth: []
  /transactions/online/{userId}:
    get:
      tags:
        - Employee
      summary: Get all online transactions initiated by the user
      description: This can only be done by the employee.
      operationId: getOnlineTransactionsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of user which initialized the transactions
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

        "400":
          description: Invalid status value.
        "404":
          description: User not found
        '403':
          description: User must be of role employee
        '401':
          description: User is not authorized
      security:
        - bearerAuth: []
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Transaction:
      type: object
      required:
        - id
        - fromAccount
        - toAccount
        - amount
        - date
        - userId
      properties:
        id:
          type: string
          format: uuid
          example: "223e4567-e89b-12d3-a456-426614174000"
        fromAccount:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        toAccount:
          type: string
          format: uuid
          example: "321e4567-e89b-12d3-a456-426614174000"
        amount:
          type: number
          format: double
          example: 250.00
        date:
          type: string
          format: date-time
          example: "2023-10-05T14:48:00.000Z"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
...